+++ src/commands/fun/randomnumber.js
@@ -33,15 +33,11 @@ module.exports = class CMD extends Command {
         const max = parseInt(args[0]) || 100;
         //TODO: Add translation
         const embed = new Embed().setDescription(
-            `Your random number between 0 and ${max} is: **${this.randomNum(
+            `Your random number between 0 and ${max} is: **${this.client.util.randomNum(
                 0,
                 max
             )}**`
         );
         message.channel.send({ embeds: [embed] });
-    }
-
-    randomNum(min, max) {
-        return Math.floor(Math.random() * (max - min)) + min;
     }
 };